feat: Initial deployment scheduling (#121)
Adds initial deployment scheduling.

This PR adds a new scheduled job which periodically scans for "pending"
jobs, and notifies the `DeploymentProvider` with the number of jobs and
the number of running machines.

The `DeploymentProvider` is then responsible for making a determination
around starting up new instances if necessary. Currently, this just
starts an instance if none are running.

This is a naive approach, we will want to move this scheduling behaviour
into a seperate service or at least introduce some locking so that only
one instance of the control plane is handling scheduling at once.

There is also no mechanism for scaling down, as this is not necessary
for `LambdaProvider` which will timeout on idle.

Please note: `deployment_id` is not currently being set on the
`machines` entities. I will create a seperate PR which set's this value
if applicable so that we can determine the running instance count for a
deployment.
Initially, I was going to fall back to checking CloudWatch metrics (for
`LambdaProvider`), however the 1 minute frequency wouldn't be granular
enough.