feat: Service registration discovery (#140)
Update the deployment build behaviour to evaluate the project and find
all service registrations by walking each (`require`'ed) file's AST with
`acorn`.

- Lambda entry-point now calls `listen()` on the service directly (as we
know where it is) rather than relying on the user's entry point to start
the service it self.
- We no longer need to run `delete require.cache[]` on each Lambda start
as the entry point will call `listen()` to restart polling (this will
allow us to leverage re-used execution context).
- We will fall back to `package.json#main` when looking for an entry
point to evaluate the project from.
- `differential deploy create` will only allow deployment of a service
found in the project. It will present a list for selection if none is
provided.

To start with, the service registration must be exported as a `default`
export.